+++ CHANGELOG.md
@@ -362,3 +362,4 @@
 
 # 0.3.8
 * Define `twosome` method to lift two parts to `<:.:> (:*:)` scheme
+* Remove `branches` method in favor of `Morphable` instance